Hawaii - Enabling of front and rear camera.
[Defect / Enhancement / New Feature]
New feature.

[Problem]
Front and Rear Camera is not working properly on HawaiiRay/Stone.

[Solution]
The patch uses regulator controls to be used by camera client to turn-on and off
the regulators when camera is opened and closed.
The controls have been integrated with the SOC camera power functions.

Note: The regulator MMCLDO1 alone is kept always ON as it seems to interfere
with the AF of primary camera. MMCLDO1 is ON by default now
Bit will be fixed and checked-in at a later point.

[Reviewers]
Vishwas, Shameer.

[Testing]
    [Test Cases]
	Camera, Camcorder w/ zoom
    [Label/CL# Synced To]
	sdb-common-android-jb
    [Platform]
	Hawaii
[BROADCOM INTERNAL NOTES]

Change-Id: Ieafee24021993aadc749380d49b96c3aa094fefe
Signed-off-by: Vinay Krishnamurthy <vinayk@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/5113
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>