arm: arch_timer: move core to drivers/clocksource
The core functionality of the arch_timer driver is not directly tied to
anything under arch/arm, and can be split out.

This patch factors out the core of the arch_timer driver, so it can be
shared with other architectures. A couple of functions are added so
that architecture-specific code can interact with the driver without
needing to touch its internals.

The ARM_ARCH_TIMER config variable is moved out to
drivers/clocksource/Kconfig, existing uses in arch/arm are replaced with
HAVE_ARM_ARCH_TIMER, which selects it.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Acked-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Hemanth Puranik <hemanth.puranik@broadcom.com>
Signed-off-by: Shankar B <sbrahade@broadcom.com>
Change-Id: I25b01b291eb1f88c3777b55f285f1396f500fd58
Reviewed-on: http://mps-gerrit.broadcom.com/26906
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Commit Description Check <commit_desc_noreply@broadcom.com>
Reviewed-by: Srivathsa Parthasarathi <parthasa@broadcom.com>
Reviewed-by: JH (Janghyun) Baek <jhbaek@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>