arm: arch_timer: divorce from local_timer api
Currently, the arch_timer driver is tied to the arm port, as it relies
on code in arch/arm/smp.c to setup and teardown timers as cores are
hotplugged on and off. The timer is registered through an arm-specific
registration mechanism, preventing sharing the driver with the arm64
port.

This patch moves the driver to using a cpu notifier instead, making it
easier to port.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Acked-by: Marc Zyngier <marc.zyngier@arm.com>
Acked-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Signed-off-by: Hemanth Puranik <hemanth.puranik@broadcom.com>
Signed-off-by: Shankar B <sbrahade@broadcom.com>
Change-Id: I20bb52c7a052714ade7fb39c03a4397589185f56
Reviewed-on: http://mps-gerrit.broadcom.com/26903
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Commit Description Check <commit_desc_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Srivathsa Parthasarathi <parthasa@broadcom.com>
Reviewed-by: JH (Janghyun) Baek <jhbaek@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>