arm: arch_timer: add arch_counter_set_user_access
Several bits in CNTKCTL reset to 0, including PL0VTEN. For architectures
using the generic timer which wish to have a fast gettimeofday vDSO
implementation, these bits must be set to 1 by the kernel. For
architectures without a vDSO, it's best to leave the bits set to 0 for
now to ensure that if and when support is added, it's implemented sanely
architecture wide.

As the bootloader might set PL0VTEN to a value that doesn't correspond
to that which the kernel prefers, we must explicitly set it to the
architecture port's preferred value.

This patch adds arch_counter_set_user_access, which sets the PL0 access
permissions to that required by the architecture. For arch/arm, this
currently means disabling all userspace access.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Hemanth Puranik <hemanth.puranik@broadcom.com>
Signed-off-by: Shankar B <sbrahade@broadcom.com>
Change-Id: I0170ac5787d53b77e50a985329f57fb5e8dcb473
Reviewed-on: http://mps-gerrit.broadcom.com/26905
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Commit Description Check <commit_desc_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Srivathsa Parthasarathi <parthasa@broadcom.com>
Reviewed-by: JH (Janghyun) Baek <jhbaek@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>