m4u: Expose the xfifo writepointer through debugfs
[Defect / Enhancement / New Feature]
Enhancement

[Problem]
If any MM access was blocked by M4U, there are no indications
to user.

[Solution]
M4U generates an error interrupt if a mapping is invalid etc.
In the ISR, the error cause is logged into a circular memory
buffer. The write pointer of the buffer is exposed via debugfs.
A non-zero value of this indicates something wrong had happened.
Using JTAG, can connect and see the buffer to see which master
caused the error and the reason.

[Reviewers]
Vinayak, Vikram

[Testing]
    [Test Cases]
    Manually removed few mapping from the PT and verified.

    [Label/CL# Synced To]
    sdb-common-android-jb

    [Platform]
    hawaii_android_edn01x

[BROADCOM INTERNAL NOTES]

Signed-off-by: Nishanth Peethambaran <nishanth@broadcom.com>
Change-Id: I2ea1f46d7979410b2b310342f1af201fa3c503c3
Reviewed-on: http://mps-gerrit.broadcom.com/7201
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Manoj Arvind <manoja@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>