m4u: Use bigger page-sizes when possible
[Defect / Enhancement / New Feature]
Enhancement

[Problem]
The page size is hard-coded to 4KB and so lot of page thrashing
happens.

[Solution]
Check the alignment of the pa and mma. Select the best possible
page size possible. To simplify the mapping needed for 8 PTE per
TLB entry, the alignment is checked for 8*pagesize.

[Reviewers]
Vinayak, Vikram

[Testing]
    [Test Cases]
    Verified from the debugfs mapping values shown

    [Label/CL# Synced To]
    sdb-common-android-jb

    [Platform]
    hawaii_android_edn01x

[BROADCOM INTERNAL NOTES]

Signed-off-by: Nishanth Peethambaran <nishanth@broadcom.com>
Change-Id: I6fabb907dfa74e0a665de65d95efdcba06f8d15e
Reviewed-on: http://mps-gerrit.broadcom.com/7205
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Manoj Arvind <manoja@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>