Make the virtual address start of each IO group alligned at 1MB.
[Defect / Enhancement / New Feature]
Defect

[Problem]
When the initial minimal page table is created in
__create_page_tables of head.S, the uart va passed
gets alligned at 1MB. But as we pass a non 1MB
alligned VA, this leads to a crash when DEBUG_LL
is enabled. This 1MB allignment is not mandated once
the mappings are created later in iotable_init. But
we have to place at least the uart VA properly
alligned to satisfy DEBUG_LL.

[Solution]
Allign the uart VA.

[Testing]
[Test Cases]
Android boot

[Label/CL# Synced To]

[Platform]
Hawaii*

[BROADCOM INTERNAL NOTES]

Change-Id: Ic7db07417ec64be640003c128f4af960783102c5
Signed-off-by: Vinayak Menon <Vinayak.Menon@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/7249
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>