ioremap: Adding a new set of APIs which will help to perform memory mapping in contexts where sleep is not allowed.
[Defect / Enhancement / New Feature]
Enhancement

[Problem]
ioremap should not be called from interrupt contexts,
but there are certain scenarios where we have to do this.
For eg.
1) We are in an ISR
2) We get the physical address of a region of RAM used by
another processor, say CP which provides the region where
it has saved the crash summary. This PA is available to us
only when we are inside the ISR, which means we cant pre-map
the area.
3) We have to map it. We use ioremap and we hit BUG!!.

[Solution]
1) Allocate VM areas during init.
2) map this preallocated VM area inside the ISR.

Currently the partitioning of allocated area, and guard region
maintainance etc are performed by the drivers calling the APIs.
This makes things simpler. But can get complicated and redundancy
will pop up when more drivers use the APIs. Currently the whole
stuff is done to rescue a single driver. So the driver does the
above said maintainance. In future we may have to move the
activities inside the APIs.

[Testing]
[Test Cases]
Any crash in interrupt context used to result in this
BUG. So do an Alt B + C and wait for u-boot to stop
at fastboot.

[Label/CL# Synced To]

[Platform]
Hawaii*

[BROADCOM INTERNAL NOTES]

Change-Id: Iea98bc4ef554f659fe8bade5f985594806039d11
Signed-off-by: Vinayak Menon <Vinayak.Menon@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/7603
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Jason (JaeSup) Shim <jshim@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>