ipc:crash support: Replacing the ioremaps used in ProcessCPCrashedDump and related functions with plat_ioremap_ns.
[Defect / Enhancement / New Feature]
Defect

[Problem]
AP gets stuck and doesnt get reset for the user
to take a ramdump in the case of a crash from
interrupt context.

[Solution]
Use plat_ioremap_ns and related functions instead
of ioremap_nocache.

The function BCMLOG_HandleCpCrashMemDumpData was
mapping 32MB of CP area in a single ioremap. But
we cant afford that as we are pre allocating vm
area. This has been split to into remaps of smaller
chunks.

[Testing]
[Test Cases]
Created a crash from interrupt context. And waited
for the device to restart and stop at u-boot
for ramdump download.

[Label/CL# Synced To]

[Platform]
Hawaii*

[BROADCOM INTERNAL NOTES]

Change-Id: I4972fc4dc98c8624d67dd4a68e44365178a10ac6
Signed-off-by: Vinayak Menon <Vinayak.Menon@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/7604
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Jason (JaeSup) Shim <jshim@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>