gpu: ion: Clarify variable names and comments around heap ids v types
There is some confusion between when to use the heap type and when
the id.  This patch clarifies this by using clearer variable names
and describing the intention in the comments.  Also fixes the client
debug code to print heaps by id instead of type.

Change-Id: Ie8b3dadded52e18590fcb2ca94001f6ed46ef07d
Signed-off-by: Rebecca Schultz Zavin <rebecca@android.com>
Signed-off-by: Nishanth Peethambaran <nishanth@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/9336
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Manoj Arvind <manoja@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>