arm:traps: Remove the code which was clearing the interrupt status to workaround the crash in interrupt context issue.
[Defect / Enhancement / New Feature]
Defect

[Problem]
We had the problem of crash inside the panic handler,
when the panic handler is called in interrupt context.
The fix is in place. But the workaround done earlier
is not removed.

[Solution]
The correct fix for the problem was done
by the commit 6a80fa0ea273bdb398862df891490a09453fb976.
So remove the workaround.

[Testing]
[Test Cases]
Android boot -> crash the kernel -> Make sure the device
restarts and stops at u-boot for ramdump download.

[Label/CL# Synced To]

[Platform]
Hawaii*

[BROADCOM INTERNAL NOTES]

Change-Id: Ib93011026f0f7969c1d3d729f36d246cf885712e
Signed-off-by: Vinayak Menon <Vinayak.Menon@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/8639
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>