V3D: Samurai NULL dereference
[Defect / Enhancement / New Feature]
MobC00265929

[Problem]
Binning allocation causes a NULL dereference as the "out of binning (overspill) memory"
interrupt is raised, but the job completes almost immediately (the current specification
says that the job will be stalled).
GFXH-26 describes a manifestation of this bug, but we also see it when a job exactly uses
the binning overspill memory. In this case, it does not need more to complete.

[Solution]
Modified the binning memory allocation to not be job-specific and to free the allocated memory
if there is no need for it by the time the allocation has completed.

Change-Id: Id327296de626433756e7e6a37d1be75c5b32608d
Signed-off-by: Roger Nixon <rnixon@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/9009
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: CQ Status <cq_status_noreply@broadcom.com>
Reviewed-by: Seetharam Samptur <ssamptur@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/9044