drivers:sd: Wait for the device resume to complete in cd ISR.
[Defect / Enhancement / New Feature]
Defect

[Problem]
When we perform a card insert/removal during a
suspend, it is possible (and it happens) that
the device resume is not complete when the cd ISR executes.
The controller interrupts are enabled only in resume.
But before the resume completes and we get a sdhci irq, if
we perform another insert/removal, the previous one is
lost forever and results in peri clock not being disabled.
And this results in passing incomplete information on card
status to upper layer.

[Solution]
Wait for the device to be resumed in cd ISR.

[Testing]
[Test Cases]
1) Clock state
2) Deep sleep
3) SD insert/removal

[Label/CL# Synced To]

[Platform]
Hawaii*

[BROADCOM INTERNAL NOTES]

Change-Id: I9fd663a8e952ad2d481e9e7c3ec03c1d7b085a18
Signed-off-by: Vinayak Menon <Vinayak.Menon@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/9455
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Naga Chandra Prahalada Rao <nagachandra@broadcom.com>
Reviewed-by: Sandeepa Prabhu <sprabhu@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>