drivers:sd: Make a direct clock enable in card emulate, to avoid wait for resume completion.
[Defect / Enhancement / New Feature]
Defect

[Problem]
The card emulate can be executed just
after a device resume. pm core enables the
runtime pm for the device when the resume
callback returns. It is possible that we
execute card emulate before that. Thus
pm_runtime_get_sync will fail to enable the
clock.

[Solution]
One option is to wait for the RPM to be enabled,
but this can create an obvious issue when we
perform card insert/removal during the probe.
We cant remove the pm_runtime_get_sync with a
direct clock enable here, because the
sdhci_irq will complain about "interrupt
while runtime suspened". We cant even check
for pm_runtime status and take an appropriate
action, because if the status changes by the
time the clcok disable is done in card tasklet,
it will result in unbalanced RPM usage. Moreover
why to wait for something in an ISR when we have
an option not to.
So a better option is to do both (get_sync and
direct clock enable).

[Testing]
[Test Cases]
1) Clock state
2) Deep sleep
3) SD insert/removal

[Label/CL# Synced To]

[Platform]
Hawaii*

[BROADCOM INTERNAL NOTES]

Change-Id: I488a8c12163b5a3db7aff55d82867122c2737ba5
Signed-off-by: Vinayak Menon <Vinayak.Menon@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/9459
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Naga Chandra Prahalada Rao <nagachandra@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>