drivers:mmc: Increase the RPM autosuspend timeout to 200ms.
[Defect / Enhancement / New Feature]
Enhancement

[Problem]
The sdhci request statistics shows that during
the write of comparatively huge file (10MB or
more) with a low performance SD card,
the time between consecutive requests is
on an average 190ms. It will be advantageous to
increase the RPM autosspend delay to a value
more than 190ms.

[Solution]
Increase the RPM autosuspend delay to 190ms.
As a drawback, this means that there will be a
worst case delay of 190ms in entering deep sleep.

[Testing]
[Test Cases]
1) Clock state
2) Deep sleep
3) SD insert/removal
4) SD throughput

The SD write with a Transcend 4GB HS card shows
an improvement of 0.2MB/s.

[Label/CL# Synced To]

[Platform]
Hawaii*

[BROADCOM INTERNAL NOTES]

Change-Id: I9c9af6faf739fd44db2cd1e8e2f60b51f645f49a
Signed-off-by: Vinayak Menon <Vinayak.Menon@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/9527
Reviewed-by: Sandeepa Prabhu <sprabhu@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>