Kona_timer: Kona hub timer gets stuck
[Defect / Enhancement / New Feature]
Defect

[Problem]
When running some of the Unit testcases to run one shot timer multiple
timers. On running this test case we would occasionally see freeze in
channel 2 or channel 3 (which are used as local timers for system)

Also, when running monkey tests we see the same symptom.

[Solution]
__disable_channel function is currently reading the STCS register,
disabling the relveant channel and writing the value back to STCS
register. If another channel had interrupt pending in STCS register
while we read the STCS register we would write back 1 to wrong channel
and miss the interrupt for that channel.

The correct method to disable channel would be
1) Read STCS
2) Mask lower 4 bits.
3) Set the relvant channel bit for TIMER_MATCH.
4) Set the relvant channel bit for COMPARE_ENABLE.
5) Writeback to STCS register.

[Reviewers]
Vikram, Shankar B.

[Testing]
    [Test Cases]
	1) Unit test to run multiple channels (which would start/stop
	   individual channels) in oneshot mode.

    [Label/CL# Synced To]

    [Platform]
	Hawaii

[BROADCOM INTERNAL NOTES]

Signed-off-by: Nagachandra P <nagachandra@broadcom.com>
Change-Id: I73f3c196d48f23f76ed59e804f1123a3ebb93639
Reviewed-on: http://mps-gerrit.broadcom.com/9766
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Commit Description Check <commit_desc_noreply@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>