mm: Modify the free page state early during set_migratetype_isolate.
[Defect / Enhancement / New Feature]
Defect

[Problem]
The free pages are accounted, but
not at the proper time. Currently the isolated
pages are accounted during a free, and not when
the pageblock is marked as isolate.

[Solution]
This patch is based on the following
mainline commits.

commit d1ce749a0db12202b711d1aba1d29e823034648d
cma: count free CMA pages
commit 2139cbe627b8910ded55148f87ee10f7485408ed
cma: fix counting of isolated page

[Label/CL# Synced To]

[Platform]
Hawaii*

[BROADCOM INTERNAL NOTES]

Change-Id: Ic1a0a9d61e22553ed61da69c1ea82be28a07e63f
Signed-off-by: Vinayak Menon <Vinayak.Menon@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/10549
Reviewed-by: Sandeep Patil <spatil@broadcom.com>
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>