Ensure SDXC_LDO is not turned off when the Debug Bus is enabled on SDDAT pins
[Defect / Enhancement / New Feature]
MobC00272603

[Problem]
Ensure SDXC_LDO is not turned off when the Debug Bus is enabled on SDDAT
pins.

There is currently a corner case where we start with no SD card (the
regulators are not turned on), and then on first suspend we end up
disabling the regulator if debug bus has turned it on (even though we
have not turned it on), as regulators are shared.

[Solution]
is_regulator_enable is replaced with a check to see if we have enabled
regulator or not. In case SD has not enabled the regulator, then do not
disable it.

[Reviewers]
Vikram MP, Shameer P, Vinayak M

[Testing]
    [Test Cases]
	1) Enable debug using
	   echo 1 1 3 f > /sys/kernel/debug/power_mgr/pm_debug_bus
	2) Boot with no SD card.
	3) Enter suspend and ensure regulators are still enabled.

    [Label/CL# Synced To]

    [Platform]
	Hawaii

[BROADCOM INTERNAL NOTES]

Signed-off-by: Nagachandra P <nagachandra@broadcom.com>
Change-Id: If4a3bb9540a2756f31acd0c545bbf63b42019d77
Reviewed-on: http://mps-gerrit.broadcom.com/10687
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Vinayak Menon <vinayak.menon@broadcom.com>
Reviewed-by: Shameer Puthalan <shameer@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>