Display: Restructure code
[Defect / Enhancement / New Feature]

[Problem]
Any bug fix that is done for a display driver has
to be applied to all the remaining driver files.

[Solution]
Restructure the code such that the display driver
specific information is stored in a separate file
and the common logic remains in a separate file.

Porting document is placed here:
http://confluence.broadcom.com/download/attachments/102368681/LCD_Porting_Guide.pdf

[Reviewers]
Vishwas PS, Vinay K, Manoj A, Haipeng, Kevin

[Testing]
    [Test Cases]
    Sanity on NT35516 and OTM1281A

    [Label/CL# Synced To]
    sdb-common-android-jb-4.2

    [Platform]
    Hawaii

[BROADCOM INTERNAL NOTES]

Change-Id: I6ccc6de4192dbbc4baad9287ffff40a8f99ea13b
Signed-off-by: Raghu M J <raghumj@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/11126
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Vishwas PS <viswasps@broadcom.com>
Reviewed-by: Manoj Arvind <manoja@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>