Display: Restructure code
[Defect / Enhancement / New Feature]

[Problem]
Any bug fix that is done for a display driver has
to be applied to all the remaining driver files.

[Solution]
Restructure the code such that the display driver
specific information is stored in a separate file
and the common logic remains in a separate file.

Porting document is placed here:
http://confluence.broadcom.com/download/attachments/102368681/LCD_Porting_Guide.pdf

[Reviewers]
Manoj A

[Testing]
[Test Cases]
Sanity on NT35516 and OTM1281A

[Label/CL# Synced To]
sdb-common-android-jb-4.2

[Platform]
Hawaii

[BROADCOM INTERNAL NOTES]

Change-Id: I3607b209799d02aa2d46eb4a1f25cb1327fabacf
Signed-off-by: Raghu M J <raghumj@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/11601
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Manoj Arvind <manoja@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>