lib/rwsem-spinlock.c - Add tracker code
[Enhancement]

[Problem]
In Earlier projects we have seen deadlocks where the task
is waiting for mm->mmap_sem and we don't know who holds this.

[Solution]
If we happen to know the owner of the semaphore that can give us clue on why
there is a deadlock. This code adds a field in the rw semaphore structure and
updates it with the task struct pointer who owns it.

[Reviewers]
Kernel Team

[Testing]
    [Test Cases]
Just boot up the phone and ensured that the new field is added.
Since the functionality is already tested in Rhea did not do functional
verificaiton.

    [Label/CL# Synced To]

    [Platform]

[BROADCOM INTERNAL NOTES]

Change-Id: I4171fcac2ac0a2df286fe55eb30721096aeede4a
Reviewed-on: http://mps-gerrit.broadcom.com/12690
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>