ashmem.c - Fix get_name, set_name not to hold the ashmem_mutex
[Defect]
MobC00282385

[Problem]
There exists a path in ashmem driver that could lead to acquistion of certain
 locks in reverse order.
 sys_mmap_pgoff (1. acquires mm->mmap_sem)
  |
   --> mmap_region
 	|
         ----> ashmem_mmap (2. acquires asmem_mutex)
 Now if there is a context switch between 1 & 2
 And if another thread (that shares the mm struct) invokes an ioctl say
 ASHMEM_GET_NAME, this can lead to the following path

ashmem_ioctl
  |
  -->get_name (1. acquires ashmem_mutex)
	|
	---> copy_to_user (2. acquires the mm->mmap_sem)
The copy_to_user could lead to a valid fault if the user space has allocated
the memory but if the physical page for the same is not mapped yet.

[Solution]
Do not call copy_to_user or copy_from_user while holding the ahsmem_mutex.
Instead copy this to a local buffer that lives in the stack while holding
this lock. This will maintain data integrity as well never reverse the lock
order.

[Reviewers]
Nishanth, Sandeep Patil, Kernel Team

[Testing]
    [Test Cases]
    Used a Hawaii Stone EDN002 HW.
    Deviced a unit test code to reproduce this problem.
    With this test case we could hit the problem 100%

    After the mentioned change ran the tests for almost 30 minutes and the
    issue was not seen.

    The same patch on Rheastone was subjected to one round of ST as well.

    [Label/CL# Synced To]

    [Platform]
    Hawaii stone

[BROADCOM INTERNAL NOTES]

Change-Id: Ib78013250ba483dbed6064833c80c99a5378ab7b
Reviewed-on: http://mps-gerrit.broadcom.com/12693
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>