block: Enabing CFQ as default IO scheduler.
[Defect / Enhancement / New Feature]
MobC00258485,MobC00281853, MobC00275432,
MobC00281749

[Problem]
At present, we use the noop IO scheduler,
but tests show that it doesnt work well
for our IO workloads. noop works well for
non seeky media like flash, but as it is
just a FIFO, it cant prioritize IO requests
to different medium on the same system from
different processes. This results in starvation.
An easy to reproduce scenario is to copy a
large file from SD card to eMMC using console
or adb, and parallely try to navigate through
the UI and try to play something. We will
notice that the UI is extremely sluggish and
almost unusable. The fact that even a filemap fault
triggers an eMMC access explains the scenario.

[Solution]
Replace noop with CFQ. CFQ provides different
IO queues for each process and provides IO
priority and timeslice. The problem explained above
is not observed with CFQ.

[Tests]
MM/SD concurrency tests shows improvement.
The SD throughput is comparable to noop.
Playback is now possible during a transfer
from SD to eMMC.

[Platform]
Hawaii*

[BROADCOM INTERNAL NOTES]

Change-Id: I07b2ef015b747272059ed81faf9e8d9ae60f736d
Signed-off-by: Vinayak Menon <Vinayak.Menon@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/12950
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Commit Description Check <commit_desc_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>