Porting the cd-rom patch from capri to hawaii.
[New Feature]
None.

[Problem]
To simulate a cd-rom, the pcs are not allowed to
write files on it.

[Solution]
If we mount an iso but not a partition,
we set it to readonly.
This change is ported from capri.
The commit id in capri is:
cc1f96db21ac3fb0f72622127f23c7dd035f7415
I remove the printk and fix the checkpatch.

[Reviewers]

[Testing]
    [Test Cases]
    Compile only. Shen has tested it.

    [Label/CL# Synced To]
    Head of sdb-common-android-jb-4.2-hybrid

    [Platform]
    Hawaii

[BROADCOM INTERNAL NOTES]

Change-Id: I5b8ea4c75059df8f5c2f73ba78d81f6a3778af79
Reviewed-on: http://mps-gerrit.broadcom.com/14409
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Commit Description Check <commit_desc_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Shen Liu <liushen@broadcom.com>
Reviewed-by: Naga Chandra Prahalada Rao <nagachandra@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>