mmc: enable pm runtime for wlan
[Defect / Enhancement / New Feature]
MobC00288117

[Problem]
Throughput issues for wifi because of the
time consumed in clk_enable.

[Solution]
This patch
1) Enables pm runtime for wifi.
2) makes a change in the pm idle callback
to make sure runtime suspend is not called
as a fallback during idle.
3) removes the funtion stop_clk and all its
references. This funtion was called without
a corresponding enable. Moreover this is not
required at all. PM runtime will take care
of the clocks.
4) mmc_detect_change need not be called in
the wifi detect routine. It will be called in
the sdhci_tasklet_card. Calling this twice can
cause a lockdep warning (or in other words a
possible deadlock).
5) pm runtime autosuspend delay for wlan is
kept as 250 msec and for others 200ms.

[Platform]
Hawaii*

[Testing]
1) WLAN ON/OFF
2) WLAN throughput
3) SD insert/removal
4) deep sleep.

[BROADCOM INTERNAL NOTES]

Change-Id: I73d776a95b4ba013f3380a2126aa4e942b4a3057
Signed-off-by: Vinayak Menon <Vinayak.Menon@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/15075
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Commit Description Check <commit_desc_noreply@broadcom.com>
Reviewed-by: CQ Status <cq_status_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>